frontend/requirements.txt
streamlit==1.42.0
requests==2.32.3
pillow==11.1.0
